sql numeric gets to translated to decimal (?)

+++ MoneyStats/MoneyStats.DAL/Model/TransactionTagConnModel.cs
@@ -2,9 +2,9 @@
 {
     public class TransactionTagConnModel
     {
-        public int Id { get; set; }
-        public int TransactionId { get; set; }
-        public int TagId { get; set; }
+        public decimal Id { get; set; }
+        public decimal TransactionId { get; set; }
+        public decimal TagId { get; set; }
 
         public TransactionModel Transaction { get; set; }
         public TagModel Tag { get; set; }
